Refactoring Server.ts
